refactor(runtime): extract snapshot-store tests into dedicated module (#3207)
