+++ docs/foundation/runtime-network.md
@@ -57,6 +57,9 @@ This document captures the initial runtime-network foundation slice for peer lif
   - `InMemoryRuntimeSnapshotStore`
   - `FileRuntimeSnapshotStore`
   - `SnapshotStoreError`
+- Added dedicated runtime snapshot-store test ownership module:
+  - `crates/kamn-core/src/runtime_tests_snapshot_store.rs`
+  - routed via path-based test module declaration from `crates/kamn-core/src/runtime_tests.rs`
 - Added watchdog anomaly and deterministic fault simulation primitives in `crates/kamn-core/src/runtime_transport_coordination.rs`:
   - `WatchdogAnomalyKind`
   - `WatchdogAnomalySeverity`
@@ -70,6 +73,14 @@ This document captures the initial runtime-network foundation slice for peer lif
   - `DeterministicNetworkFaultSimulator`
   - `simulate_daemon_network_fault(...)`
 - Kept runtime role wiring behavior unchanged and covered by existing tests.
+
+## Runtime Test Module Ownership Rules
+- Snapshot-store test cases are owned by `crates/kamn-core/src/runtime_tests_snapshot_store.rs`.
+- `crates/kamn-core/src/runtime_tests.rs` retains routing via:
+  - `#[path = "runtime_tests_snapshot_store.rs"]`
+  - `mod runtime_tests_snapshot_store;`
+- Regression contract:
+  - snapshot-store test definitions are not re-inlined into `runtime_tests.rs` (`Regression: #3207`).
 
 ## Node CLI Recovery-Check Mapping
 - `kamn-node --runtime-mode recovery-check` maps directly to `RecoveryRejoinGuard` evaluation flow.
